Get most of the context switch working
We can now successfully switch context out of the kernel,
and into a user "task" (just a function pointer for now).
The current problem is getting the linker to store a branch
instruction in 0x8, so that SWI jumps to the kernel exception
handler.
Therefore, at the moment, we can't re-enter the kernel when
the task is finished running.