Don't inline syscalls
I can't convince the compiler that the return value from the syscall in
r0 actually exists, and therefore isn't something it can stomp.
By putting the syscall in a separate compilation unit, this forces the
compiler not to make assumptions about whether it can stomp r0 or not.
Adding the extra layer of indirection to solve a compiler issue seems
stupid, and I would like to revisit this later.