Add switchable buffered / busy-wait IO backend
Currently, we just use the busy-wait backend, since
we don't care about response times, and we shouldn't
be manipulating kernel buffers from user tasks.

This allows us to get rid of the duplicated code in bwio.o,
which is agnostic to the actual output method, and is just
text processing (e.g.: printf).