Refactor of both context switches
It turns out that there's this very useful variant of `msr`, which
allows us to manipulate the cpsr mode & interrupt bits directly.
In addition to requiring 1 instruction instead of 3 for mode switches,
this doesn't require a scratch register.
This lets us do a mode switch immediately without saving a register
on the stack.

This makes our SWI and IRQ interrupt handlers similar enough that they
can be factored together.
We now have a macro which has a bit of conditional logic in it,
and can expand to either handler.