Initial prototype of interrupt driven IO
It currently is only set up for COM1, but we just need to turn
interrupts on for COM2 in order for it to work for COM2 as well.
Also, this works when outputting data via COM1 to a terminal, but since
we don't do anything with the CTS bit right now, this won't
transmit data to the trains properly.

Also, this isn't very well tested, and I'm not super confident that the
tx_pending and rx_pending paths are getting exercised right now.