Simplify state machine for interrupt IO
We used to ACK TX & RX interrupts after one comes in when no one is
listening (by turning off TX/RX interrupts until a task wants the
result).
Now, we just always turn off TX/RX interrupts when no task is listening.
This avoids handling an interrupt when we know nobody wants to know
about it.
It also allows us to strip out a bunch of code, which is always nice.