Get debug output to actually work
It turns out that we don't seem to be able to get output to work for
COM3 on the TS7200. I still really have no idea why this is, but I'm
deeply confused by what I read about COM3 in the documentation.
The TS7200 documentation barely mentions it, and the Cirrus
documentation claims that UART2 (the one we're using for COM2) is for an
infrared device.