WIP on trying to get the sensors to not freak out
Right now, they sometimes choke on garbage input from the trainset, and
produce the wrong sensor numbers.
This appears to be because a single byte is left over from a sensor poll
from a previous run.
Even when we clear this out, it still messes up, and we're really not
sure why.
I'm going to write code to display past sensor reads in order to better
debug this.