Switch IO function args order to have channel at end
This is the convention used by fputs & fgets family of functions.
We used to not follow this convention, and GCC would assume that we had
defined them in the standard way, and do interesting optimizations.
The alternative to complying with the standard would be to turn off GCC
warnings about printf & friends altogether.